--- a/gulpfile.js
+++ b/gulpfile.js
@@ -1282,6 +1282,14 @@ process.on('beforeExit', function() {
 });
 
 
-gulp.on('task_start', (e) => { analytics.buildStart('gulp ' + e.task)});
-gulp.on('task_stop', (e) => { analytics.buildSuccess('gulp ' + e.task, e.duration*1000)});
-gulp.on('task_err', (e) => { analytics.buildError('gulp ' + e.task, e.duration*1000)});
+var firstTask = true;
+gulp.on('task_start', (e) => {
+  if (firstTask) {
+    firstTask = false;
+    analytics.buildSuccess('gulp <startup>', process.uptime() * 1000);
+  }
+
+  analytics.buildStart('gulp ' + e.task)
+});
+gulp.on('task_stop', (e) => {analytics.buildSuccess('gulp ' + e.task, e.duration * 1000)});
+gulp.on('task_err', (e) => {analytics.buildError('gulp ' + e.task, e.duration * 1000)});