--- a/tools/analytics/analytics.js
+++ b/tools/analytics/analytics.js
@@ -2,6 +2,7 @@
 
 let execSync = require('child_process').execSync;
 let fs = require('fs');
+let minimist = require('minimist');
 let path = require('path');
 let os = require('os');
 let ua;
@@ -58,6 +59,8 @@ let customParams = {
   cd8: `${os.cpus().length} x ${os.cpus()[0].model}`,
   // HW - Memory Info
   cd9: `${Math.round(os.totalmem()/1024/1024/1024)}GB`,
+  // gulp --projects (angular2,angular2_material)
+  cd13: minimist(process.argv.slice(2)).projects
 };