@IgorMinar 2015-11-13
I haven't adjusted the node and dart trees yet. Let's see if this passes on CI and if it does let's merge it. I'll redo the other two trees tomorrow in as followup PR.

note that I called the flag `--projects`. this is because I want to rename modules to projects (or just wait until modules goes away when @yjbanov's directory layout is implemented).