@IgorMinar 2015-11-13
We can improve the perf further by dropping the ES6 emit when `--projects` is specified (or removing it completely as proposed by @alexeagle)