@IgorMinar 2015-11-13
for casual testing it looks like `--projects=angular2` speeds up the build as follows:

|                 | build time | test time |
-------------- | --------------|-------------
| **first build**  | 21s -> 17s | 14s -> 12s |
| **rebuild**     | 4.5s -> 3.8s | 11s -> 10s |


We are still spending too much time in Funnels, so there is more we can do to optimize this.

Before:

```
Slowest Trees                                 | Total               
----------------------------------------------+---------------------
DiffingTSCompiler                             | 2285ms              
Funnel                                        | 753ms               
DiffingTSCompiler                             | 321ms               

Slowest Trees (cumulative)                    | Total (avg)         
----------------------------------------------+---------------------
DiffingTSCompiler (2)                         | 2606ms (1303 ms)    
Funnel (85)                                   | 1349ms (15 ms)      
MergeTrees (5)                                | 393ms (78 ms)       
```


After:

```
Slowest Trees                                 | Total               
----------------------------------------------+---------------------
DiffingTSCompiler                             | 1977ms              
Funnel                                        | 846ms               

Slowest Trees (cumulative)                    | Total (avg)         
----------------------------------------------+---------------------
DiffingTSCompiler (2)                         | 2092ms (1046 ms)    
Funnel (6)                                    | 1042ms (173 ms)     
MergeTrees (5)                                | 376ms (75 ms) 
```


Dropping the ES6 output is likely the most impactful thing we can do next.