@mary-poppins 2015-11-13
Merging PR #5272 on behalf of @vsavkin to branch [presubmit-vsavkin-pr-5272](https://github.com/angular/angular/tree/presubmit-vsavkin-pr-5272).